Add a lot of keys to mapping table

+++ sdl_keys.h
@@ -4,5 +4,5 @@
 #define SDL_KEYS 323
 #define  VK_KEYS 256
 
-extern Uint8 VKey_to_SDLKey[VK_KEYS];
-extern int SDLKey_to_VKey[SDL_KEYS];
+extern int VKey_to_SDLKey[VK_KEYS];
+extern Uint8 SDLKey_to_VKey[SDL_KEYS];
